Add files via upload
When using Folder Locker, the user sets up a unique password, which serves as the key to access the locked folders. Once the password is set, the user can then select the folders they wish to protect and initiate the locking process. After locking, the selected folders become inaccessible to anyone who doesn't know the password.